OPENNLP-264 Code Cleanup: Remove unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/incubator/opennlp/trunk@1161001 13f79535-47bb-0310-9956-ffa450edef68

+++ opennlp-tools/src/main/java/opennlp/tools/sentdetect/SentenceDetectorME.java
@@ -244,7 +244,7 @@ public class SentenceDetectorME implements SentenceDetector {
   public double[] getSentenceProbabilities() {
     double[] sentProbArray = new double[sentProbs.size()];
     for (int i = 0; i < sentProbArray.length; i++) {
-      sentProbArray[i] = ((Double) sentProbs.get(i)).doubleValue();
+      sentProbArray[i] = sentProbs.get(i).doubleValue();
     }
     return sentProbArray;
   }
